Moved tests to the 'src/tests' folder to match other testing groups.
