--- oom-tests/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "lib": ["es2020"],
-    "types": ["node", "jest"],
-    "noEmit": true
-  },
-  "include": [
-    "../src/**/*",
-    "./**/*.ts"
-  ]
-}
-
